Build sucessful. moose.py is __init__.py _moose.so|pydis moose.so|pyd These changes are required for successful import on NT. Tests are passing. Not all but most.
