+++ python/moose/fixXreacs.py
@@ -14,15 +14,15 @@ from __future__ import print_function, division, absolute_import
 ####################################################################
 
 import sys
-import moose._moose as _moose
+import moose.moose as cppmoose
 
 msgSeparator = "_xMsg_"
 
 def findCompt( elm ):
-    elm = _moose.element( elm )
+    elm = cppmoose.element( elm )
     pa = elm.parent
     while pa.path != '/':
-        if _moose.Neutral(pa).isA[ 'ChemCompt' ]:
+        if cppmoose.Neutral(pa).isA[ 'ChemCompt' ]:
             return pa.path
         pa = pa.parent
     print( 'Error: No compartment parent found for ' + elm.path )
@@ -33,7 +33,7 @@ def checkEqual(lst):
     return not lst or lst.count(lst[0]) == len(lst)
 
 def findXreacs( basepath, reacType ):
-    reacs = _moose.wildcardFind( basepath + '/##[ISA=' + reacType + 'Base]' )
+    reacs = cppmoose.wildcardFind( basepath + '/##[ISA=' + reacType + 'Base]' )
     ret = []
     for i in reacs:
         reacc = findCompt( i )
@@ -50,38 +50,38 @@ def findXreacs( basepath, reacType ):
     return ret
 
 def removeEnzFromPool( pool ):
-    kids = _moose.wildcardFind( pool.path + "/#" )
+    kids = cppmoose.wildcardFind( pool.path + "/#" )
     for i in kids:
         if i.isA[ 'EnzBase' ]:
-            _moose.delete( i )
+            cppmoose.delete( i )
         elif i.isA[ 'Function' ]:
-            _moose.delete( i )
+            cppmoose.delete( i )
 
 # If a pool is not in the same compt as reac, make a proxy in the reac 
 # compt, connect it up, and disconnect the one in the old compt.
 def proxify( reac, reacc, direction, pool, poolc ):
-    reacc_elm = _moose.element( reacc )
-    reac_elm = _moose.element( reac )
+    reacc_elm = cppmoose.element( reacc )
+    reac_elm = cppmoose.element( reac )
     # Preserve the rates which were set up for the x-compt reacn
-    #_moose.showfield( reac )
-    dupname = pool.name + '_xfer_' + _moose.element(poolc).name
+    #cppmoose.showfield( reac )
+    dupname = pool.name + '_xfer_' + cppmoose.element(poolc).name
     #print "#############", pool, dupname, poolc
-    if _moose.exists( reacc + '/' + dupname ):
-        duppool = _moose.element( reacc + '/' + dupname )
+    if cppmoose.exists( reacc + '/' + dupname ):
+        duppool = cppmoose.element( reacc + '/' + dupname )
     else:
         # This also deals with cases where the duppool is buffered.
-        duppool = _moose.copy(pool, reacc_elm, dupname )
+        duppool = cppmoose.copy(pool, reacc_elm, dupname )
     duppool.diffConst = 0   # diffusion only happens in original compt
     removeEnzFromPool( duppool )
     disconnectReactant( reac, pool, duppool )
-    _moose.connect( reac, direction, duppool, 'reac' )
-    #_moose.showfield( reac )
-    #_moose.showmsg( reac )
+    cppmoose.connect( reac, direction, duppool, 'reac' )
+    #cppmoose.showfield( reac )
+    #cppmoose.showmsg( reac )
 
 def enzProxify( enz, enzc, direction, pool, poolc ):
     if enzc == poolc:
         return
-    enze = _moose.element( enz )
+    enze = cppmoose.element( enz )
     # kcat and k2 are indept of volume, just time^-1
     km = enze.numKm
     proxify( enz, enzc, direction, pool, poolc )
@@ -90,7 +90,7 @@ def enzProxify( enz, enzc, direction, pool, poolc ):
 def reacProxify( reac, reacc, direction, pool, poolc ):
     if reacc == poolc:
         return
-    reac_elm = _moose.element( reac )
+    reac_elm = cppmoose.element( reac )
     kf = reac_elm.numKf
     kb = reac_elm.numKb
     proxify( reac, reacc, direction, pool, poolc )
@@ -108,26 +108,26 @@ def identifyMsg( src, srcOut, dest ):
 def disconnectReactant( reacOrEnz, reactant, duppool ):
     outMsgs = reacOrEnz.msgOut
     infoPath = duppool.path + '/info'
-    if _moose.exists( infoPath ):
-        info = _moose.element( infoPath )
+    if cppmoose.exists( infoPath ):
+        info = cppmoose.element( infoPath )
     else:
-        info = _moose.Annotator( infoPath )
+        info = cppmoose.Annotator( infoPath )
 
-    #_moose.le( reactant )
+    #cppmoose.le( reactant )
     notes = ""
-    #_moose.showmsg( reacOrEnz )
+    #cppmoose.showmsg( reacOrEnz )
     for i in outMsgs:
         #print "killing msg from {} to {}\nfor {} and {}".format( reacOrEnz.path, reactant.path, i.srcFieldsOnE1[0], i.srcFieldsOnE2[0] )
         if i.e1 == reactant:
             msgStr = identifyMsg( i.e2, i.e2.srcFieldsOnE2[0], i.e1 )
             if len( msgStr ) > 0:
                 notes += msgStr
-                _moose.delete( i )
+                cppmoose.delete( i )
         elif i.e2 == reactant:
             msgStr = identifyMsg( i.e1[0], i.srcFieldsOnE1[0], i.e2[0] )
             if len( msgStr ) > 0:
                 notes += msgStr
-                _moose.delete( i )
+                cppmoose.delete( i )
     #print "MSGS to rebuild:", notes
     info.notes += notes
 
@@ -154,7 +154,7 @@ def fixXreacs( basepath ):
 def getOldRates( msgs ):
     if len( msgs ) > 1 :
         m1 = msgs[1].split( msgSeparator )[0]
-        elm = _moose.element( m1.split( ' ' )[0] )
+        elm = cppmoose.element( m1.split( ' ' )[0] )
         if elm.isA[ 'ReacBase' ]:
             return [elm.numKf, elm.numKb]
         elif elm.isA[ 'EnzBase' ]:
@@ -166,7 +166,7 @@ def restoreOldRates( oldRates, msgs ):
     #print oldRates, msgs
     if len( msgs ) > 1 :
         m1 = msgs[1].split( msgSeparator )[0]
-        elm = _moose.element( m1.split( ' ' )[0] )
+        elm = cppmoose.element( m1.split( ' ' )[0] )
         if elm.isA[ 'ReacBase' ]:
             elm.numKf = oldRates[0]
             elm.numKb = oldRates[1]
@@ -176,20 +176,20 @@ def restoreOldRates( oldRates, msgs ):
 
 
 def restoreXreacs( basepath ):
-    proxyInfo = _moose.wildcardFind( basepath + "/##/#_xfer_#/info" )
+    proxyInfo = cppmoose.wildcardFind( basepath + "/##/#_xfer_#/info" )
     for i in proxyInfo:
         msgs = i.notes.split( msgSeparator )
         oldRates = getOldRates( msgs )
         #print( "Deleting {}".format( i.parent.path ) )
         #print msgs
-        _moose.delete( i.parent )
+        cppmoose.delete( i.parent )
         for j in msgs[1:]:
             if len( j ) > 0:
                 args = j.split( ' ' )
                 assert( len( args ) == 4 )
-                #_moose.showfield( args[0] )
-                _moose.connect( args[0], args[1], args[2], args[3] )
+                #cppmoose.showfield( args[0] )
+                cppmoose.connect( args[0], args[1], args[2], args[3] )
                 #print( "Reconnecting {}".format( args ) )
-                #_moose.showfield( args[0] )
+                #cppmoose.showfield( args[0] )
         restoreOldRates( oldRates, msgs )
