+++ tests/python/test_moose_attribs.py
@@ -38,7 +38,7 @@ attribs = ['AdExIF', 'AdThreshIF', 'Adaptor', 'Annotator', 'Arith', 'BufPool',
         'Table2', 'TableBase', 'TimeTable', 'VClamp', 'VERSION', 'Variable',
         'VectorTable', 'ZombieBufPool', 'ZombieCaConc', 'ZombieCompartment',
         'ZombieEnz', 'ZombieFunction', 'ZombieHHChannel', 'ZombieMMenz',
-        'ZombiePool', 'ZombieReac', '_moose', 
+        'ZombiePool', 'ZombieReac', 'moose', 
         'append_finfodocs', 'ce', 'chemMerge',
         'chemUtil', 'closing', 'connect', 'copy', 'delete', 'division', 'doc',
         'element', 'exists', 'finfotypes', 'fixXreacs', 'genesis', 'getCwe',
